references/agentic-framework-main/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlmodel>=0.0.14
alembic>=1.13.0

# Storage & Databases
redis>=5.0.0
psycopg2-binary>=2.9.9
pymilvus>=2.3.0
chromadb>=0.4.18
minio>=7.2.0

# Task Queue
celery>=5.3.0

# ML & Embeddings
sentence-transformers>=2.2.2

# Document Q&A Skill
haystack-ai>=2.0.0
markitdown>=0.1.0
Pillow>=10.0.0

# Document Processing
python-docx>=1.1.0
pypdf>=4.0.0
openpyxl>=3.1.0
python-pptx>=0.6.23
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# LLM Providers
anthropic>=0.18.0
tiktoken>=0.5.0

# Auth & Security
authlib>=1.3.0
python-jose[cryptography]>=3.3.0

# Observability
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
prometheus-client>=0.19.0

# HTTP & Async
httpx>=0.25.0
python-multipart>=0.0.6
anyio>=4.0.0

# Schema Validation
jsonschema>=4.20.0

# CLI Tool (Kautilya)
click>=8.1.7
rich>=13.7.0
prompt-toolkit>=3.0.43
questionary>=2.0.1
jinja2>=3.1.2
python-dotenv>=1.0.0
openai>=1.0.0
ddgs>=9.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
mypy>=1.7.0
ruff>=0.1.6
